+++ qoqo/src/operations/mod.rs
@@ -89,6 +89,7 @@ use pyo3::prelude::*;
 ///    PragmaRandomNoise
 ///    PragmaGeneralNoise
 ///    PragmaConditional
+///    PragmaLoop
 ///    CNOT
 ///    SWAP
 ///    FSwap
@@ -158,6 +159,7 @@ pub fn operations(_py: Python, m: &PyModule) -> PyResult<()> {
     m.add_class::<PragmaGeneralNoiseWrapper>()?;
     m.add_class::<PragmaConditionalWrapper>()?;
     m.add_class::<PragmaChangeDeviceWrapper>()?;
+    m.add_class::<PragmaLoopWrapper>()?;
     m.add_class::<CNOTWrapper>()?;
     m.add_class::<SWAPWrapper>()?;
     m.add_class::<FSwapWrapper>()?;